+++ src/components/auth/signup/step2/style.scss
@@ -1,4 +1,4 @@
-.email-verification-btn {
+.phone-verification-btn {
     border: none;
     margin-bottom: 10px;
     margin-left: 230px;
