+++ src/components/auth/signup/step3/index.tsx
@@ -1,51 +1,31 @@
-import TextField from "@/components/common/textField";
-import useSignup from "@/hooks/auth/useSignup";
-import { Toast } from "@/libs/toast";
-import { verificationEmailCode } from "@/api/verification.api";
-import './style.scss';
+import TextField from '@/components/common/textField';
+import useSignup from '@/hooks/auth/useSignup';
 
-const Step2 = () => {
-    const { signupData, onChange, onKeyDown } = useSignup();
-
-    const handleEmailVerification = async () => {
-        try {
-            await verificationEmailCode(signupData.email);
-            Toast("success", "이메일 코드 전송 완료");
-        } catch (error) {
-            Toast("error", "이메일 코드 전송 실패");
-        };
-    };
+const Step4 = () => {
+    const { signupData, onChange } = useSignup();
 
     return (
-        <div className='step2-container'>
-            <div className='step2-input-group'>
+        <div className='step4-container'>
+            <div className='step4-input-group'>
                 <TextField 
-                    label='전화번호'
-                    name='phone'
-                    placeholder='전화번호를 입력해주세요'
-                    value={signupData.phone}
-                    onChange={onChange}
-                    onKeyDown={onKeyDown}
-                />
-            </div>
-            <div className='step2-input-group'>
-                <TextField 
-                    label='인증번호'
-                    name='emailVerificationCode'
-                    placeholder='인증번호를 입력해주세요'
-                    value={signupData.emailVerificationCode}
+                    label='소속'
+                    name='userClass'
+                    placeholder='소속을 입력해주세요'
+                    value={signupData.userClass}
                     onChange={onChange}
                 />
             </div>
-            <button
-                type="button"
-                onClick={handleEmailVerification}
-                className="phone-verification-btn"
-            >
-                인증번호 재요청
-            </button>
+            <div className='step4-input-group'>
+                <TextField 
+                    label='직군'
+                    name='field'
+                    placeholder='직군을 입력해주세요'
+                    value={signupData.field}
+                    onChange={onChange}
+                />
+            </div>
         </div>
     );
 };
 
-export default Step2;
+export default Step4;
