--- src/components/auth/signup/step3/style.scss
@@ -1,7 +1,0 @@
-.phone-verification-btn {
-    border: none;
-    margin-bottom: 10px;
-    margin-left: 230px;
-    background-color: var(--Common-00);
-    color: var(--Primary-Normal);
-}
