--- src/components/common/oauthButton/style.scss
@@ -1,26 +1,0 @@
-.oauth-button-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-top: 30px;
-    margin-bottom: 30px;
-}
-
-.oauth-button {
-    display: flex;
-    gap: 20px;
-}
-
-.google, .kakao {
-    width: 140px;
-    height: 50px;
-    border-radius: 14px;
-    border: none;
-    background-color: var(--Common-00);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
-    cursor: pointer;
-    transition: all 0.3s ease;
-}
