+++ src/components/common/progressbar/index.tsx
@@ -1,0 +1,9 @@
+import './style.scss';
+
+const ProgressBar = () => {
+    return (
+        <div></div>
+    );
+};
+
+export default ProgressBar;
