+++ src/pages/login/style.scss
@@ -34,7 +34,6 @@
 }
 
 .logo-container {
-    margin-bottom: 30px;
     margin-left: -10px;
 }
