+++ src/pages/signup/page.tsx
@@ -5,11 +5,10 @@ import useSignup from '@/hooks/auth/useSignup';
 import Logo from '@/assets/logo.svg';
 import SignupBanner from '@/assets/signupBanner.svg';
 import Button from '@/components/common/button/index';
-import OauthButton from '@/components/common/oauthButton/index';
 import Step1 from '@/components/auth/signup/step1';
 import Step2 from '@/components/auth/signup/step2';
-import Step3 from '@/components/auth/signup/step3';
-import Step4 from '@/components/auth/signup/step4';
+import Step3 from '@/components/auth/signup/step2';
+import Step4 from '@/components/auth/signup/step3';
 import './style.scss';
 
 const Signup = () => {
@@ -22,7 +21,7 @@ const Signup = () => {
     };
 
     const handleNextStep = () => {
-        if (step < 4) {
+        if (step < 3) {
             setStep(step + 1);
         } else {
             handleSignup();
@@ -49,7 +48,6 @@ const Signup = () => {
                         <div className='signup-button-group'>
                             <Button text={ step < 4 ? '다음' : '가입하기' } onClick={ handleNextStep } />
                         </div>
-                        <OauthButton />
                         <div className='login-redirect'>
                             이미 계정이 있으신가요? 
                             <span onClick={ handleLoginRedirect }> 로그인 하기</span>
